backend/requirements.txt
Django
djangorestframework
celery
redis
django-celery-beat
django-celery-results